[OC-1123] avoids infinity loop on unfound 'respond to' entities of a card
In the a case of a card with an existing list of
`entitiesAllowedToRespond`, if the referenced entities are not found
then only one round of looking up is done. Otherwise an infinity loop is
performed.
Thus, now a check is done to ensure that the entities to respond for the
current card has been initialized. If so, no look up is performed.

Signed-off-by: LE-GALL Ronan Ext <43667786+rlg-rte@users.noreply.github.com>

lints some ts files

Signed-off-by: LE-GALL Ronan Ext <43667786+rlg-rte@users.noreply.github.com>

[OC-1123]

Signed-off-by: LE-GALL Ronan Ext <43667786+rlg-rte@users.noreply.github.com>