Refactor Sound Activation to Use New Modal Service
This commit transitions the sound activation feature to use the new modal service.
This eliminates the need for the separate Angular component previously used to open the sound activation popup.
As a result, the 'sound-activation' Angular component has been removed, and the modal is now directly opened from the sound notification service.

Signed-off-by: freddidierRTE <frederic-f.didier@rte-france.com>