Authorize special characters in processInstanceId (#8361)
Add base64 encoding in id in urls on the angular side and configure spring to allowed special
characters.

The only characters that is not allowed is the slash as it causes
problem in spring when parsing the url path.

A cypress card has been updated with a processInstanceId containing
special characters so to be able to detect potential future regressions

Signed-off-by: freddidierRTE <frederic-f.didier@rte-france.com>